backend/requirements.txt
# Level Backend Requirements

# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ML and embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Clustering and dimensionality reduction
scikit-learn>=1.3.0
umap-learn>=0.5.0
scipy>=1.11.0

# Gemini API
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Optional: For future multimodal support
# pillow>=10.0.0
# librosa>=0.10.0
# opencv-python>=4.8.0
